# frntEnd/openapi.yaml
openapi: 3.1.0
info:
  title: AI Workflow Builder API
  version: 1.0.0
  description: |
    REST API for the AI Workflow Builder platform.

    ## Authentication
    All endpoints (except `/auth/*` and `/health`) require a valid JWT Bearer token
    or an API Key in the `X-API-Key` header.

    ## Tenant Isolation
    Every request is scoped to a tenant. The tenant is resolved from the JWT claim
    (`tenant_id`) or from the API key row. Cross-tenant access is impossible by design.

    ## Error Envelope
    All error responses use the standard envelope below:
    ```json
    {
      "error": {
        "code": "WORKFLOW_NOT_FOUND",
        "message": "Workflow wf_abc123 not found.",
        "request_id": "req_01J..."
      }
    }
    ```

  contact:
    name: Platform Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.workflowplatform.io/v1
    description: Production (us-east-1)
  - url: https://api.eu.workflowplatform.io/v1
    description: EU Production (eu-west-1 — v1.2+)
  - url: http://localhost:8000/v1
    description: Local development

# ─────────────────────────────────────────────────────────────────────────────
# Security Schemes
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed JWT. Obtain via `/auth/login` or `/auth/token/refresh`.

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Tenant API key (prefix `wfk_`). Created in the dashboard.
        Scopes are checked per-endpoint.

  # ───────────────────────────────────────────────────────────────────────────
  # Reusable Schemas
  # ───────────────────────────────────────────────────────────────────────────
  schemas:
...
    ErrorDetail:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: WORKFLOW_NOT_FOUND
        message:
          type: string
          example: "Workflow wf_abc123 not found."
        request_id:
          type: string
          example: req_01JRXYZ
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'

    # ── Pagination ─────────────────────────────────────────────────────────────
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        has_next:
          type: boolean

    # ── Auth ───────────────────────────────────────────────────────────────────
...
    WorkflowMetadata:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        tags:
          type: array
          items:
            type: string
          maxItems: 20
        is_active:
          type: boolean
          default: true
    NodePosition:
      type: object
      required: [x, y]
      properties:
        x:
          type: number
          description: Horizontal canvas position in pixels.
        y:
          type: number
          description: Vertical canvas position in pixels.
    NodeUIConfig:
      type: object
      description: |
        Frontend rendering hints. Populated by WorkflowLayoutEngine on DAG generation.
        Frontend reads these to style and label each node; never used by execution engine.
      properties:
        editable:
          type: boolean
          default: true
          description: Whether the node's config can be edited in the UI panel.
        node_type_label:
          type: string
          description: Human-readable display label for the node type badge.
          example: "AI Prompt"
        icon:
          type: string
          description: Icon identifier (maps to frontend icon registry).
          example: "sparkles"
        color:
          type: string
          description: Hex colour for the node header band.
          example: "#6366f1"
        category:
          type: string
          enum: [ai_reasoning, execution_data, workflow_management, logic_orchestration, triggers]
          description: Node category for palette grouping.
        is_terminal:
          type: boolean
          default: false
          description: True for output/trigger nodes — prevents deletion in UI.
    WorkflowUIMetadata:
      type: object
      description: Frontend-specific workflow metadata. Not used by execution engine.
      properties:
        layout:
          type: string
          enum: [auto, manual]
          default: auto
          description: |
            auto — positions were set by WorkflowLayoutEngine.
            manual — user has moved nodes; preserve exact positions.
        version:
          type: string
          default: "1.0"
        viewport:
          type: object
          description: Last saved React Flow viewport (pan + zoom).
          properties:
            x: { type: number, default: 0 }
            y: { type: number, default: 0 }
            zoom: { type: number, default: 1 }
        generated_by_chat:
          type: boolean
          default: false
          description: True when this workflow was created via the chat DAG generator.
        chat_session_id:
          type: string
          nullable: true
          description: Session ID of the generating chat conversation, if applicable.
    WorkflowDefinition:
      type: object
      description: |
        Graph definition. nodes is a dict keyed by node_id.
        edges connect output ports to input ports.
        ui_metadata carries React Flow viewport state and layout hints.
      required: [nodes, edges]
      properties:
        nodes:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/NodeDefinition'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/EdgeDefinition'
        ui_metadata:
          $ref: '#/components/schemas/WorkflowUIMetadata'
    NodeDefinition:
      type: object
      required: [type, config]
      properties:
        type:
          type: string
          description: |
            Node type identifier. Each workflow must have exactly one trigger node
            (manual_trigger, scheduled_trigger, or integration_trigger) as its root.
          enum:
            # AI & Reasoning
            - prompt
            - agent
            - semantic_search
            # Execution & Data
            - code_execution
            - api_request
            - templating
            - web_search
            - mcp
            # Workflow Management
            - set_state
            - custom
            - note
            - output
            # Logic & Orchestration
            - control_flow
            - subworkflow
            # Triggers
            - manual_trigger
            - scheduled_trigger
            - integration_trigger
        config:
          type: object
          description: Node-type-specific configuration object. Schema varies by type.
        position:
          $ref: '#/components/schemas/NodePosition'
        ui_config:
          $ref: '#/components/schemas/NodeUIConfig'
    EdgeDefinition:
      type: object
      required: [id, source_node_id, source_port, target_node_id, target_port]
      properties:
        id:
          type: string
          description: Unique edge identifier (used by React Flow as edge key).
          example: "edge_node1_output__node2_input"
        source_node_id:
          type: string
        source_port:
          type: string
          description: Output port name on source node (e.g. "output", "true", "false").
        target_node_id:
          type: string
        target_port:
          type: string
          description: Input port name on target node (e.g. "input").
    WorkflowSummary:
      type: object
      required: [workflow_id, tenant_id, metadata, version_no, created_at, updated_at]
      properties:
        workflow_id:
          type: string
        tenant_id:
          type: string
          format: uuid
        metadata:
          $ref: '#/components/schemas/WorkflowMetadata'
        version_no:
          type: integer
          description: Latest published version number.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkflowDetail:
      allOf:
        - $ref: '#/components/schemas/WorkflowSummary'
        - type: object
          required: [definition]
          properties:
            definition:
              $ref: '#/components/schemas/WorkflowDefinition'
    WorkflowCreate:
      type: object
      required: [definition, metadata]
      properties:
        definition:
          $ref: '#/components/schemas/WorkflowDefinition'
        metadata:
          $ref: '#/components/schemas/WorkflowMetadata'
    WorkflowUpdate:
      type: object
      properties:
        definition:
          $ref: '#/components/schemas/WorkflowDefinition'
        metadata:
          $ref: '#/components/schemas/WorkflowMetadata'
        publish:
          type: boolean
          description: If true, saves a new version snapshot.
          default: false

    # ── Workflow Versions ──────────────────────────────────────────────────────
    WorkflowVersionSummary:
      type: object
      required: [version_id, workflow_id, version_no, sdk_version, created_by, created_at]
      properties:
        version_id:
          type: string
        workflow_id:
          type: string
        version_no:
          type: integer
        sdk_version:
          type: string
          example: "1.0.0"
        change_summary:
          type: string
          nullable: true
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    WorkflowVersionDetail:
      allOf:
        - $ref: '#/components/schemas/WorkflowVersionSummary'
        - type: object
          required: [definition]
          properties:
            definition:
              $ref: '#/components/schemas/WorkflowDefinition'

    # ── Executions ─────────────────────────────────────────────────────────────
...
    ScheduleCreate:
      type: object
      required: [cron_expression, timezone]
      properties:
        cron_expression:
          type: string
          description: Standard 5-field cron expression.
          example: "0 9 * * MON-FRI"
        timezone:
          type: string
          description: IANA timezone identifier.
          example: "America/New_York"
        trigger_input:
          type: object
          description: Static payload to inject at each fire.
        is_active:
          type: boolean
          default: true
    Schedule:
      allOf:
        - $ref: '#/components/schemas/ScheduleCreate'
        - type: object
          required: [schedule_id, workflow_id, next_fire_at, created_at]
          properties:
            schedule_id:
              type: string
            workflow_id:
              type: string
            next_fire_at:
              type: string
              format: date-time
            last_fired_at:
              type: string
              format: date-time
              nullable: true
            created_at:
              type: string
              format: date-time

    # ── Webhooks ───────────────────────────────────────────────────────────────
...
    WorkflowUpdateRequest:
      type: object
      description: Payload sent from frontend when user modifies the generated DAG.
      required: [nodes, edges]
      properties:
        workflow_id:
          type: string
          description: ID of the workflow being edited. Used for audit trail.
        updated_nodes:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/NodeDefinition'
          description: Full node map after edits (replaces entire nodes dict).
        updated_edges:
          type: array
          items:
            $ref: '#/components/schemas/EdgeDefinition'
          description: Full edge list after edits.
        ui_metadata:
          $ref: '#/components/schemas/WorkflowUIMetadata'
          description: Updated viewport + layout mode from frontend.
    WorkflowUpdateResponse:
      type: object
      description: Validation result returned after frontend submits workflow edits.
      required: [valid, workflow]
      properties:
        valid:
          type: boolean
        workflow:
          $ref: '#/components/schemas/WorkflowDefinition'
        validation_errors:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
              node_id: { type: string, nullable: true }
          description: Present only when valid is false.
        suggestions:
          type: array
          items: { type: string }
          description: Optional AI-generated improvement hints (e.g. "Consider adding an error handler after node_3").
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UNAUTHORIZED
              message: "Bearer token is missing or expired."
              request_id: req_01JRXYZ
    Forbidden:
      description: Authenticated but insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found (or not visible to this tenant).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    UnprocessableEntity:
      description: Validation error — request body failed schema or business rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    TooManyRequests:
      description: Rate limit or quota exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the limit resets.
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'

  # ───────────────────────────────────────────────────────────────────────────
  # Reusable Parameters
  # ───────────────────────────────────────────────────────────────────────────
  parameters:
    WorkflowId:
      name: workflow_id
      in: path
      required: true
      schema:
        type: string
      description: Workflow identifier (e.g. wf_01JRXYZ).
...
    VersionNo:
      name: version_no
      in: path
      required: true
      schema:
        type: integer
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

# ─────────────────────────────────────────────────────────────────────────────
# Global Security (overridable per-operation)
# ─────────────────────────────────────────────────────────────────────────────
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Tags
# ─────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /workflows:
    get:
      operationId: workflows_list
      tags: [Workflows]
      summary: List workflows for the current tenant
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
          description: Filter by workflow name (prefix match).
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [updated_at, name]
            default: updated_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Paginated workflow list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkflowSummary'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: workflows_create
      tags: [Workflows]
      summary: Create a new workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowCreate'
      responses:
        '201':
          description: Workflow created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /workflows/{workflow_id}:
    get:
      operationId: workflows_get
      tags: [Workflows]
      summary: Get workflow definition (latest draft)
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: Workflow detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: workflows_update
      tags: [Workflows]
      summary: Update workflow definition and/or metadata
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowUpdate'
      responses:
        '200':
          description: Workflow updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      operationId: workflows_delete
      tags: [Workflows]
      summary: Delete a workflow (soft delete — OWNER only)
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '204':
          description: Workflow deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workflows/{workflow_id}/activate:
    post:
      operationId: workflows_activate
      tags: [Workflows]
      summary: Activate a workflow (EDITOR+ required)
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: Workflow activated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /workflows/{workflow_id}/deactivate:
    post:
      operationId: workflows_deactivate
      tags: [Workflows]
      summary: Deactivate a workflow (stops schedules and new triggers)
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: Workflow deactivated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Workflow Versions ─────────────────────────────────────────────────────────
  /workflows/{workflow_id}/versions:
    get:
      operationId: workflow_versions_list
      tags: [Versions]
      summary: List published versions for a workflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Version list (newest first).
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkflowVersionSummary'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /workflows/{workflow_id}/versions/{version_no}:
    get:
      operationId: workflow_versions_get
      tags: [Versions]
      summary: Get a specific version snapshot (immutable)
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
        - $ref: '#/components/parameters/VersionNo'
      responses:
        '200':
          description: Version detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowVersionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /workflows/{workflow_id}/versions/{version_no}/restore:
    post:
      operationId: workflow_versions_restore
      tags: [Versions]
      summary: Restore a version as the current draft (creates new version on next publish)
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
        - $ref: '#/components/parameters/VersionNo'
      responses:
        '200':
          description: Draft restored from version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Executions ────────────────────────────────────────────────────────────────
...
  /workflows/{workflow_id}/schedules:
    get:
      operationId: schedules_list
      tags: [Schedules]
      summary: List schedules for a workflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: Schedule list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Schedule'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      operationId: schedules_create
      tags: [Schedules]
      summary: Create a cron schedule for a workflow (EDITOR+ required)
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleCreate'
      responses:
        '201':
          description: Schedule created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /chat/sessions/{session_id}/workflow:
    put:
      operationId: chat_update_workflow
      tags: [Chat]
      summary: Submit frontend workflow edits — validates DAG and returns suggestions
      description: |
        Called when the user modifies nodes or edges in the DAG preview canvas.
        Runs cycle detection + port compatibility check. On success, persists the
        updated WorkflowDefinition and returns optional AI improvement suggestions.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowUpdateRequest'
      responses:
        '200':
          description: Validation result with updated workflow and optional suggestions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowUpdateResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'